+++ src/main/java/org/openlmis/referencedata/repository/SupplyLineRepository.java
@@ -19,16 +19,15 @@ import org.javers.spring.annotation.JaversSpringDataAuditable;
 import org.openlmis.referencedata.domain.Facility;
 import org.openlmis.referencedata.domain.SupplyLine;
 import org.openlmis.referencedata.repository.custom.SupplyLineRepositoryCustom;
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 import java.util.UUID;
 
 @JaversSpringDataAuditable
-public interface SupplyLineRepository extends
-        PagingAndSortingRepository<SupplyLine, UUID>,
+public interface SupplyLineRepository extends JpaRepository<SupplyLine, UUID>,
     SupplyLineRepositoryCustom {
 
   @Query(name = "SupplyLine.findSupplyingFacilities")